Separated Operator and LazyTable into different files
